+++ src/components/SimpleModal.tsx
@@ -6,14 +6,16 @@ interface SimpleModalProps {
   isOpen: boolean;
   onRequestClose: () => void;
   children: React.ReactNode;
+  width?: string;
+  height?: string;
 }
 
-const SimpleModal: React.FC<SimpleModalProps> = ({ isOpen, onRequestClose, children }) => {
+const SimpleModal: React.FC<SimpleModalProps> = ({ isOpen, onRequestClose, children, width = '98%', height = '98%' }) => {
   if (!isOpen) return null;
 
   return (
     <div className="fixed inset-0 bg-gray-800 bg-opacity-50 flex justify-center items-center z-50">
-      <div className="bg-white rounded p-4 w-[98%] h-[98%] max-w-screen-2xl relative">
+      <div className="bg-white rounded p-4 relative" style={{ width, height }}>
         <button onClick={onRequestClose} className="absolute top-2 right-2 text-black">
           X
         </button>