+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## unreleased
 
+- Added UI for selecting filter options
+
 ## [0.1.2] - 2024-06-06
 
 - Added option to compare/diff the input and output